fuzz_trim: Move `count_expected_live` to `dfa_spec_check_liveness`.
This liveness check will also be needed for testing minimization.
