+++ theft/fuzz_trim.c
@@ -6,17 +6,11 @@
 
 #include "type_info_dfa.h"
 
-#include <adt/queue.h>
-
 static enum theft_trial_res
 check_trimming_matches_reachability(const struct dfa_spec *spec);
 
 static enum theft_trial_res
 prop_trimming_matches_reachability(struct theft *t, void *arg1);
-
-static bool
-count_expected_live(const struct dfa_spec *spec,
-    fsm_state_t start_state, size_t *exp);
 
 static bool
 test_dfa_trimming(theft_seed seed,
@@ -99,8 +93,8 @@ check_trimming_matches_reachability(const struct dfa_spec *spec)
 	}
 
 	size_t expected_post_states;
-	if (!count_expected_live(spec, spec->start, &expected_post_states)) {
-		fprintf(stderr, "-- ERROR: count_expected_live alloc\n");
+	if (!dfa_spec_check_liveness(spec, NULL, &expected_post_states)) {
+		fprintf(stderr, "-- ERROR: dfa_spec_check_liveness\n");
 		return THEFT_TRIAL_ERROR;
 	}
 
@@ -119,118 +113,6 @@ check_trimming_matches_reachability(const struct dfa_spec *spec)
 
 	fsm_free(fsm);
 	return THEFT_TRIAL_PASS;
-}
-
-static bool
-enqueue_reachable(const struct dfa_spec *spec, struct queue *q,
-    uint64_t *seen, fsm_state_t state)
-{
-	if (BITSET_CHECK(seen, state)) { return true; }
-	if (!queue_push(q, state)) { return false; }
-	BITSET_SET(seen, state);
-
-	const struct dfa_spec_state *s = &spec->states[state];
-	assert(s->used);
-
-	for (size_t i = 0; i < s->edge_count; i++) {
-		const fsm_state_t to = s->edges[i].state;
-		if (BITSET_CHECK(seen, to)) { continue; }
-		if (!enqueue_reachable(spec, q, seen, to)) {
-			return false;
-		}
-	}
-
-	return true;
-}
-
-static bool
-calc_reaches_end_by_fixpoint(const struct dfa_spec *spec,
-    uint64_t *reaches_end)
-{
-	/* A state reaches an end if it's an end, or one of its
-	 * edges transitively reaches an end. */
-	bool changed;
-	do {
-		changed = false;
-		for (size_t s_i = 0; s_i < spec->state_count; s_i++) {
-			const struct dfa_spec_state *s = &spec->states[s_i];
-			if (!s->used) { continue; }
-
-			if (s->end) {
-				if (!BITSET_CHECK(reaches_end, s_i)) {
-					changed = true;
-					BITSET_SET(reaches_end, s_i);
-				}
-				continue;
-			}
-
-			for (size_t i = 0; i < s->edge_count; i++) {
-				const fsm_state_t to = s->edges[i].state;
-				if (BITSET_CHECK(reaches_end, to)) {
-					if (!BITSET_CHECK(reaches_end, s_i)) {
-						changed = true;
-						BITSET_SET(reaches_end, s_i);
-					}
-				}
-			}
-		}
-	} while (changed);
-	return true;
-}
-
-static bool
-count_expected_live(const struct dfa_spec *spec, fsm_state_t start_state,
-	size_t *exp)
-{
-	bool res = false;
-	uint64_t *seen = NULL;
-	uint64_t *reaches_end = NULL;
-	size_t exp_count = 0;
-	struct queue *q = NULL;
-
-	if (spec->state_count == 0) {
-		*exp = 0;
-		return true;
-	}
-
-	q = queue_new(NULL, spec->state_count);
-	if (q == NULL) {
-		fprintf(stderr, "cleanup %d\n", __LINE__);
-		goto cleanup;
-	}
-
-	seen = calloc(spec->state_count/64 + 1, sizeof(seen[0]));
-	if (seen == NULL) { goto cleanup; }
-
-	reaches_end = calloc(spec->state_count/64 + 1, sizeof(reaches_end[0]));
-	if (reaches_end == NULL) { goto cleanup; }
-
-	if (!enqueue_reachable(spec, q, seen, start_state)) {
-		fprintf(stderr, "-- ERROR: enqueue_reachable\n");
-		goto cleanup;
-	}
-
-	if (!calc_reaches_end_by_fixpoint(spec, reaches_end)) {
-		fprintf(stderr, "-- ERROR: calc_reaches_end_by_fixpoint\n");
-		goto cleanup;
-	}
-
-	fsm_state_t s;
-	while (queue_pop(q, &s)) {
-		if (BITSET_CHECK(reaches_end, s)) {
-			exp_count++;
-		}
-	}
-
-	res = true;
-	*exp = exp_count;
-
-cleanup:
-	if (seen != NULL) { free(seen); }
-	if (reaches_end != NULL) { free(reaches_end); }
-	if (q != NULL) { queue_free(q); }
-
-	return res;
 }
 
 #define RUN_SPEC(STATES, START)					 \
